Firmware: Major redesing again - added OTA and modified measurements
Now boards can be updated over-the-air. Pin GP0 needs to be pulled to
GND for at least one second to enable this. Afterwards you should see
the board in the Port list of Arduino. The board reset automatically
after the update.
Sotware now sends temperature, humidity and humidity status in the same
message (last one is required by Domoticz), so there is only one sensor
needed for each board (no more separate temperature/humidity). This also
simplified the nodered config.
Also the way the measurements are taken was changed: the program takes
one measurement each 30 seconds from when it starts. After 10 minutes
and 5 seconds, it plots an average of these measurements and sends it to
the MQTT server. After a few more seconds it reboots and starts again.
The wifi is enabled only when sending the MQTT message or if the OTA is
started until the board reboots.